% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat.r
\name{as_mat}
\alias{as_mat}
\title{as_mat}
\usage{
as_mat(x, copy = TRUE)
}
\arguments{
\item{x}{R matrix.}

\item{copy}{Should the R data be copied? If \code{FALSE}, be careful!}
}
\value{
A mat object.
}
\description{
Convert an R matrix to a mat object.
}
