% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refmat.r
\name{as_refmat}
\alias{as_refmat}
\title{as_refmat}
\usage{
as_refmat(x)
}
\arguments{
\item{x}{R matrix.}

\item{copy}{Should the R data be copied? If \code{FALSE}, be careful!}
}
\value{
A refmat object.
}
\description{
Convert an R matrix to a refmat object.
}
