% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec.r
\name{as_vec}
\alias{as_vec}
\title{as_vec}
\usage{
as_vec(x, copy = TRUE)
}
\arguments{
\item{x}{R vector.}

\item{copy}{Should the R data be copied? If \code{FALSE}, be careful!}
}
\value{
A vec object.
}
\description{
Convert an R vector to a cpumat object.
}
