% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/craze-package.r
\docType{package}
\name{craze-package}
\alias{craze-package}
\title{craze}
\description{
A reference implementation of the fmlr API. The implementation only uses base
R functions.
}
\author{
Drew Schmidt
}
\keyword{package}
