% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmlref-package.r
\docType{package}
\name{fmlref-package}
\alias{fmlref-package}
\title{fmlref}
\description{
A reference implementation of the fmlr API. The implementation only uses base
R functions.
}
\author{
Drew Schmidt
}
\keyword{package}
