% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat.r
\name{mat}
\alias{mat}
\title{mat}
\usage{
mat(nrows = 0, ncols = 0, type = "double")
}
\arguments{
\item{nrows, ncols}{The dimensions of the matrix.}

\item{type}{Storage type for the matrix. Should be one of 'int', 'float', or 'double'.}
}
\value{
A mat class object.
}
\description{
Constructor for mat objects.
}
\details{
Data is held in an external pointer.
}
\seealso{
\code{\link{mat-class}}
}
