% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refvec.r
\name{refvec}
\alias{refvec}
\title{refvec}
\usage{
refvec(size = 0, type = "double")
}
\arguments{
\item{size}{Length of the vector.}

\item{type}{Storage type for the vector. Should be one of 'int' or 'double'.}
}
\value{
A refvec class object.
}
\description{
Constructor for refvec objects.
}
\details{
Data is held in an external pointer.
}
\seealso{
\code{\link{vec-class}}
}
