% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec.r
\name{vec}
\alias{vec}
\title{vec}
\usage{
vec(size = 0, type = "double")
}
\arguments{
\item{size}{Length of the vector.}

\item{type}{Storage type for the vector. Should be one of 'int', 'float', or 'double'.}
}
\value{
A vec class object.
}
\description{
Constructor for vec objects.
}
\details{
Data is held in an external pointer.
}
\seealso{
\code{\link{vec-class}}
}
